Fix 'short iPoint' bug.  Point index should always be long.

+++ SU2_CFD/include/fluctuating_stress.hpp
@@ -82,7 +82,7 @@ class CFluctuatingStress {
    */
   virtual void CalculateEddyViscosity(const CGeometry* geometry,
                                       CConfig* config,
-                                      unsigned short iPoint,
+                                      unsigned long iPoint,
                                       su2double** eddy_viscosity) const = 0;
  protected:
   const unsigned short nDim;
@@ -97,7 +97,7 @@ class CM43Model : public CFluctuatingStress {
 
   void CalculateEddyViscosity(const CGeometry* geometry,
                               CConfig* config,
-                              unsigned short iPoint,
+                              unsigned long iPoint,
                               su2double** eddy_viscosity) const;
  private:
   unsigned short density_index;
@@ -109,6 +109,6 @@ class CNoStressModel : public CFluctuatingStress {
 
   void CalculateEddyViscosity(const CGeometry* geometry,
                               CConfig* config,
-                              unsigned short iPoint,
+                              unsigned long iPoint,
                               su2double** eddy_viscosity) const;
 };
